fixed up checking for partner n stuff

+++ src/utils/events.ts
@@ -31,50 +31,53 @@ export async function intiChatClient(): Promise<void> {
 
         if (!message.startsWith(`${CONFIG.prefix}verify`)) {
             if (STORAGE.checkChat) {
+                let verified = false;
+                if (msg.userInfo.isBroadcaster) verified = true;
+                if (msg.userInfo.isVip) verified = true;
+                if (msg.userInfo.isMod) verified = true;
+                if (msg.userInfo.isSubscriber) verified = true;
+                if (msg.userInfo.badges.has("partner")) verified = true;
 
 
-                const foundUser = STORAGE.verify.find((chan) => chan.channel === user);
-                if (foundUser === undefined) {
-                    const newVerifiy: Verifiy = {
-                        attempts: 0,
-                        channel: user,
-                        hasVerified: false,
-                        id: nanoid()
-                    };
+                if (!verified) {
 
-                    STORAGE.verify.push(newVerifiy);
-                    Storage.saveConfig();
-                    chatClient.deleteMessage(channel, msg).catch(console.error);
-                    return chatClient.say(channel, `You are not verified @${msg.userInfo.displayName}`
+
+                    const foundUser = STORAGE.verify.find((chan) => chan.channel === user);
+
+                    if (foundUser === undefined) {
+                        const newVerifiy: Verifiy = {
+                            attempts: 0,
+                            channel: user,
+                            hasVerified: false,
+                            id: nanoid()
+                        };
+
+                        STORAGE.verify.push(newVerifiy);
+                        Storage.saveConfig();
+                        chatClient.deleteMessage(channel, msg).catch(console.error);
+                        return chatClient.say(channel, `You are not verified @${msg.userInfo.displayName}`
                     + ` Please type in chat ${CONFIG.prefix}verify ${newVerifiy.id}, Fail this 3 times and you'll be banned`);
 
-                }
-
-                if (!foundUser.hasVerified) {
-                    if (msg.userInfo.isBroadcaster) return;
-                    if (msg.userInfo.isVip) return;
-                    if (msg.userInfo.isMod) return;
-                    if (msg.userInfo.isSubscriber) return;
-                    if (msg.userInfo.badges.has("partner")) return;
-
-
-                    foundUser.attempts += 1;
-
-                    Storage.saveConfig();
-
-                    if (foundUser.attempts > 3) {
-                        chatClient.ban(channel, user, "Did not verify, you can still appeal though!").catch(console.error);
-                        foundUser.attempts = 0;
-                        foundUser.hasVerified = false;
-                        Storage.saveConfig();
-                        return chatClient.say(channel, `@${msg.userInfo.displayName} was banned for being unable to verify!`);
                     }
+                    if (!foundUser.hasVerified) {
+
+                        foundUser.attempts += 1;
+                        Storage.saveConfig();
+
+                        if (foundUser.attempts > 3) {
+                            chatClient.ban(channel, user, "Did not verify, you can still appeal though!").catch(console.error);
+                            foundUser.attempts = 0;
+                            foundUser.hasVerified = false;
+                            Storage.saveConfig();
+                            return chatClient.say(channel, `@${msg.userInfo.displayName} was banned for being unable to verify!`);
+                        }
 
 
-                    chatClient.deleteMessage(channel, msg).catch(console.error);
-                    return chatClient.say(channel, `You are not verified @${msg.userInfo.displayName}`
+                        chatClient.deleteMessage(channel, msg).catch(console.error);
+                        return chatClient.say(channel, `You are not verified @${msg.userInfo.displayName}`
                 + ` Please type in chat ${CONFIG.prefix}verify ${foundUser.id}, Fail this 3 times and you'll be banned`);
 
+                    }
                 }
             }
         }
